# fix(storage): simple8b passes checkptr (#15512)